rename things. start fixing lint errors
